Get it compiling

+++ src/main/scala/trw/dbsubsetter/db/Constants.scala
@@ -3,5 +3,5 @@ package trw.dbsubsetter.db
 object Constants {
 
   // Must be in descending order
-  val dataCopyBatchSizes: Seq[Int] = Seq(2048, 1024, 256, 16, 4, 2, 1)
+  val dataCopyBatchSizes: Seq[Short] = Seq(2048, 1024, 256, 16, 4, 2, 1)
 }
